WordPress & Social Login, Social Sharing Plugin Update
Simply updated WP & Social Login, Social Sharing

+++ wp-content/plugins/miniorange-login-openid/uninstall.php
@@ -1,4 +1,7 @@
 <?php
+
+	if ( !defined( 'WP_UNINSTALL_PLUGIN' ) ) 
+		exit();
 	//delete all stored key-value pairs
 
 	delete_option('mo_openid_host_name');
@@ -29,6 +32,16 @@
 	delete_option('mo_openid_login_redirect_url');
 	delete_option('moopenid_social_login_avatar');
 	delete_option('moopenid_user_attributes');
+
+	delete_option( 'mo_openid_social_comment_fb');
+	delete_option( 'mo_openid_social_comment_google' );
+	delete_option( 'mo_openid_social_comment_default' );
+	delete_option( 'mo_openid_social_comment_blogpost' );
+	delete_option( 'mo_openid_social_comment_static' );
+	delete_option('mo_openid_social_comment_default_label' );
+	delete_option('mo_openid_social_comment_fb_label' );
+	delete_option('mo_openid_social_comment_google_label' );
+	delete_option('mo_openid_social_comment_heading_label' );
 
 	delete_option('mo_share_options_enable_post');
 	delete_option('mo_share_options_enable_home_page');
@@ -62,6 +75,8 @@
 	delete_option('mo_openid_login_custom_theme' );
 	delete_option( 'mo_login_icon_custom_color');
 	delete_option('mo_share_options_enable_post_position');
+	delete_option('mo_share_options_home_page_position');
+	delete_option('mo_share_options_static_pages_position');
 	delete_option('mo_openid_share_twitter_username');
 	
 	delete_option('mo_sharing_icon_custom_size');
